backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
prisma==0.15.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
chromadb==1.5.1
sentence-transformers==3.3.1
langchain-text-splitters==0.3.2
google-genai==1.65.0
email-validator==2.2.0
